- Deleted LoadTables   - I noticed that LoadTables was an issue after deleting TableRowFilterSupport   - LoadTables has a few methods that should be called directly from Analyster   - LoadTables was actually using an instance of Analyster making it is difficult to work with.   - The UML and code on GitHub should show that It should clearly be methods in Analyster.
