+++ src/main/java/com/elle/analyster/service/DeleteRecord.java
@@ -39,7 +39,7 @@ public class DeleteRecord {
             }
             
             // this is where the table is refreshing 
-            new LoadTables().loadTable(Analyster.getInstance().getTabs().get(tableName).getTable());
+            Analyster.getInstance().loadTable(Analyster.getInstance().getTabs().get(tableName).getTable());
 
             // output pop up dialog that a record was deleted 
             JOptionPane.showMessageDialog(Analyster.getInstance(), rowsSelected + " Record(s) Deleted");
